+++ git-revision/src/spec/mod.rs
@@ -1,0 +1,20 @@
+/// How to interpret a revision specification, or `revspec`.
+#[derive(Debug, Copy, Clone, PartialOrd, PartialEq, Ord, Eq, Hash)]
+#[cfg_attr(feature = "serde1", derive(serde::Serialize, serde::Deserialize))]
+pub enum Kind {
+    /// A single revision specification, pointing at one reference.
+    Single,
+    /// Two revision specifications `a` and `b` where we want all commits from `b` that are not also in `a`.
+    Range,
+    /// Everything in `a` and `b` but no commit from any of their merge bases.
+    MergeBase,
+}
+
+impl Default for Kind {
+    fn default() -> Self {
+        Kind::Single
+    }
+}
+
+pub mod parse;
+pub use parse::function::parse;
